+++ gazebo/gui/model/ImportDialog.cc
@@ -134,7 +134,8 @@ void ImportDialog::OnCancel()
 void ImportDialog::OnImport()
 {
   QFileInfo info(this->pathLineEdit->text());
-  if (info.isFile())
+  if (info.isFile() && (info.completeSuffix().toLower() == "dae" ||
+      info.completeSuffix().toLower() == "stl"))
   {
     this->accept();
   }